api/flash: Lift out the common flashing process into api/flash
Instead of each flasher implementing roughly the same steps over and over again,
lift out the common parts (eeprom save & restore, bootloader trigger & wait,
reconnect) into a common function, and dispatch only the flashing parts to the
various helpers.
