+++ src/api/hardware-softhruf-splitography/index.js
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { DFUProgrammer } from "@api/flash/DFUProgrammerFlasher";
+import { flash, flashers } from "@api/flash";
 import Keymap from "./components/Keymap";
 
 const Splitography = {
@@ -49,7 +49,7 @@ const Splitography = {
     return ["saveEEPROM", "flash", "reconnect", "restoreEEPROM"];
   },
   flash: async (port, filename, options) => {
-    return await DFUProgrammer(null, port, filename, options);
+    return await flash(flashers.dfuProgrammer, null, port, filename, options);
   },
 };
